move Landingpage,Locationresults and SearchResults to pages
