--- src/pages/search.tsx
@@ -1,2 +1,0 @@
-import React from 'react';
-import SearchResults from 'components/SearchResults';
